util/sdnotify: only supported on !windows
