+++ pkg/cli/start_unix.go
@@ -1,0 +1,54 @@
+// Copyright 2017 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+// +build !windows
+
+package cli
+
+import (
+	"os"
+	"os/exec"
+	"strings"
+
+	"github.com/cockroachdb/cockroach/pkg/cli/cliflags"
+	"github.com/cockroachdb/cockroach/pkg/util/sdnotify"
+)
+
+var startBackground bool
+
+func init() {
+	boolFlag(startCmd.Flags(), &startBackground, cliflags.Background, false)
+}
+
+func maybeRerunBackground() (bool, error) {
+	if startBackground {
+		args := make([]string, 0, len(os.Args))
+		foundBackground := false
+		for _, arg := range os.Args {
+			if arg == "--background" || strings.HasPrefix(arg, "--background=") {
+				foundBackground = true
+				continue
+			}
+			args = append(args, arg)
+		}
+		if !foundBackground {
+			args = append(args, "--background=false")
+		}
+		cmd := exec.Command(args[0], args[1:]...)
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = stderr
+		return true, sdnotify.Exec(cmd)
+	}
+	return false, nil
+}
