+++ pkg/util/sdnotify/sdnotify_unix_test.go
@@ -1,0 +1,46 @@
+// Copyright 2016 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License. See the AUTHORS file
+// for names of contributors.
+//
+// Author: Ben Darnell
+
+// +build !windows
+
+package sdnotify
+
+import (
+	"testing"
+
+	_ "github.com/cockroachdb/cockroach/pkg/util/log" // for flags
+)
+
+func TestSDNotify(t *testing.T) {
+	l, err := listen()
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer func() { _ = l.close() }()
+
+	ch := make(chan error)
+	go func() {
+		ch <- l.wait()
+	}()
+
+	if err := notify(l.Path, readyMsg); err != nil {
+		t.Fatal(err)
+	}
+	if err := <-ch; err != nil {
+		t.Fatal(err)
+	}
+}
